+++ src/features/profile/LogoutConfirmButton.tsx
@@ -1,6 +1,6 @@
 import { Button } from '@/components/ui/button';
+import { userKeys } from '@/hooks/queries/user/userQueryKeys';
 import { logoutAuth } from '@/services/apis/authApi';
-import { userKeys } from '@/services/queryKeys';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 export default function LogoutConfirmButton() {
